+++ test/client/unit/components/app.spec.js
